incentive delete implemented

+++ frontend/src/main/java/de/uniba/dsg/dsam/client/IncentiveServlet.java
@@ -1,8 +1,10 @@
 package de.uniba.dsg.dsam.client;
 
+import de.uniba.dsg.dsam.model.Beverage;
 import de.uniba.dsg.dsam.model.Incentive;
 import de.uniba.dsg.dsam.model.PromotionalGift;
 import de.uniba.dsg.dsam.model.TrialPackage;
+import de.uniba.dsg.dsam.persistence.BeverageManagement;
 import de.uniba.dsg.dsam.persistence.IncentiveManagement;
 
 import javax.ejb.EJB;
@@ -24,6 +26,9 @@ public class IncentiveServlet extends HttpServlet {
 	
 	@EJB
 	private IncentiveManagement<?, Incentive> incentiveManager;
+
+	@EJB
+	private BeverageManagement<Beverage> beverageManagement;
 
 	/**
 	 * @see HttpServlet#doGet(HttpServletRequest request, HttpServletResponse response)
@@ -102,10 +107,32 @@ public class IncentiveServlet extends HttpServlet {
 	 * @see HttpServlet#doDelete(HttpServletRequest, HttpServletResponse)
 	 */
 	protected void doDelete(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		int id = Integer.valueOf(request.getParameter("id"));
-		List<Incentive> incentives = this.incentiveManager.getAll();
-		Incentive incentive = incentives.stream().filter(incnt -> incnt.getId() == Integer.valueOf(id)).findAny().get();
-		incentiveManager.deleteOne(incentive);
-	}
+		try{
+			int id = Integer.valueOf(request.getParameter("id"));
+			int i =0;
+			List<Incentive> incentives = this.incentiveManager.getAll();
+			List<Beverage> beverages = this.beverageManagement.getAll();
+			for(Beverage beverage : beverages){
+				if(beverage.getIncentive().isPresent()){
+					Incentive incentive_details= beverage.getIncentive().get();
+					if(incentive_details.getId()== id){
+						response.getWriter().write("Not Delete");
+						break;
+					}
+					else{
+						i=1;
 
+					}
+				}
+			}
+			if(i==1){
+				Incentive incentive = incentives.stream().filter(incnt -> incnt.getId() == Integer.valueOf(id)).findAny().get();
+				incentiveManager.deleteOne(incentive);
+				response.getWriter().write("Delete");
+			}
+		}
+		catch (NumberFormatException e){
+			System.out.println("Invalid ID "+ e.getMessage());
+		}
+	}
 }
